6: Added new cache types
Added 4 new cache types: Podcast Cache, Educational Cache, Challenge Cache, Guest Book Cache.

Outstanding issues:
The map2.php has not been tested more than that the new check boxes are displayed because
I'm unable to get the map to work on IE and on Firefox the vertical scrollbar doesn't
work properly making the screen be clipped. But no caches at all was chown...

To the Podcast Cache type is not complete in that the user should be able to upload
an mp3, this will be handled in Issue 7.

Signed-off-by: OlofL <olof_laurin@yahoo.com>