Use date_created instead of last_modified when reading latest logs for the RSS-feed
The last modified date of the log is updated when the cache's status is
changed (due to details in the XML integration interface). Since last_modified
does not always reflect the modification date date_created is used instead.
That way old logs won't turn up in the RSS-feed when the cache is archived.