Resize to big uploaded pictures
Automatic resizing of uploaded pictures it their size (in bytes) is above a threashould (settings).
The new size is determined from settings in pixels. This solution isn't optimal for
bmp since they will be much too large, but another issue would be to convert
bmp format to something more useful for the web.